+++ client/src/content/pages/Addfriend.js
