+++ client/src/content/pages/Profile.js
@@ -1,0 +1,35 @@
+import React from 'react';
+
+function Profile(props) {
+    return (
+        <div className="profile">
+            <div className='friendlist'>
+                <div>
+                    
+                </div>
+                <div>
+
+                </div>
+                <div>
+
+                </div>
+
+
+            </div>
+            <div className='events'>
+                <div>
+                    
+                </div>
+                <div>
+    
+                </div>
+                <div>
+    
+                </div>
+            </div>
+        </div>
+
+    )
+}
+
+export default Profile;
