--- eatvite/src/content/pages/Profile.js
@@ -1,21 +1,0 @@
-import React from 'react';
-
-function Profile(props) {
-    return (
-        <div className="profile">
-            <div className='friendlist'>
-                <h2>Friends</h2>
-                
-
-            </div>
-            <div className='events'>
-                <h2>Events</h2>
-
-
-            </div>
-        </div>
-
-    )
-}
-
-export default Profile;
